requirements.txt
pdfminer.six
pymupdf
fitz
pyrogram
tgcrypto